+++ src/js/kit/animate.js
@@ -12,7 +12,7 @@ function animate(el, attr, from, to, duration){
       animate(el, attr, from, to, duration - 10);
     }, 10);
   } else { throw { message: "Invalid parameters passed to css()" }; }
-};
+}
 
 // function animate(el, attr, to, duration){
 //   if (duration < 0) {return;}